! cdi: Falscher Typ in storage.h korrigiert
git-svn-id: svn+ssh://overgames.de/lost/trunk@687 1fb02b30-9e10-0410-89f7-8f5a202ca6a9

+++ include/cdi/storage.h
@@ -18,7 +18,7 @@
 struct cdi_storage_device {
     struct cdi_device   dev;
     size_t              block_size;
-    size_t              block_count;
+    uint64_t            block_count;
 };
 
 struct cdi_storage_driver {
